+++ src/apps/shared/index.ts
@@ -1,5 +1,5 @@
-import { path } from '@tauri-apps/api';
-import { getRootElement, invoke, SeelenCommand } from 'seelen-core';
+import { invoke } from '@tauri-apps/api/core';
+import { getRootElement, SeelenCommand } from 'seelen-core';
 
 export const getRootContainer = getRootElement;
 
@@ -7,11 +7,11 @@ export function toPhysicalPixels(size: number): number {
   return Math.round(size * window.devicePixelRatio);
 }
 
-export async function wasInstalledUsingMSIX() {
-  let installPath = await path.resourceDir();
-  return installPath.startsWith('C:\\Program Files\\WindowsApps');
+export async function wasInstalledUsingMSIX(): Promise<boolean> {
+  // Todo replace this when added to SeelenCommand
+  return invoke('is_appx_package');
 }
 
-export async function isDev() {
+export async function isDev(): Promise<boolean> {
   return invoke(SeelenCommand.IsDevMode);
 }