+++ src/background/utils/winver.rs
@@ -15,11 +15,18 @@ pub fn is_virtual_desktop_supported() -> bool {
     false
 }
 
-pub fn is_msix_intallation() -> bool {
+pub fn is_running_as_appx_package() -> bool {
     unsafe {
         let mut len = 0u32;
         let _ = GetCurrentPackageId(&mut len, None);
         let mut buffer = vec![0u8; len as usize];
         GetCurrentPackageId(&mut len, Some(buffer.as_mut_ptr())).is_ok()
     }
+}
+
+pub fn was_installed_using_msix() -> bool {
+    std::env::current_exe().is_ok_and(|p| {
+        p.with_file_name("AppxManifest.xml").exists()
+            || p.starts_with("C:\\Program Files\\WindowsApps")
+    })
 }