+++ changelog.md
@@ -5,6 +5,9 @@
 - add new setting on dock and toolbar to maintain overlap state by monitor.
 - add service to restart the seelen-ui app on crash.
 
+### enhancements
+- force run the app as an APPX if it was installed using msix.
+
 ### refactor
 - custom http server to serve files instead bundle it in the executable on development (local).